Updating has_artifact field in ADS
